+++ raspberrypi/aarch64/python2.7/acrcloud/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/python                                                                                                                                             
+#-*- coding: utf-8 -*-
