+++ raspberrypi/aarch64/python3/acrcloud/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/python                                                                                                                                             
+#-*- coding: utf-8 -*-
+
+import os,sys
+
+sys.path.append(os.path.split(os.path.realpath(__file__))[0])
