--- raspberrypi/python2.7/acrcloud/__init__.py
@@ -1,2 +1,0 @@
-#!/usr/bin/python                                                                                                                                             
-#-*- coding: utf-8 -*-
